REFACTOR main/404 + main/footer
