+++ src/plugins/main/footer/Footer.tsx
