+++ src/plugins/main/footer/PrivacyPolicyLink.tsx
